+++ src/ServiceFabric/ServiceFabric/ChangeLog.md
@@ -18,6 +18,7 @@
         - Additional information about change #1
 -->
 ## Upcoming Release
+* Added cmdlet `Get-AzServiceFabricRuntimeSupportedVersion` to enumerate supported runtime service fabric cluster versions in ARM. similar to cmdlet `Get-ServiceFabricRuntimeSupportedVersion` used for standalone.
 
 ## Version 2.4.0
 * Upgraded Managed Cluster commands to use Service Fabric Managed Cluster SDK version 1.0.0 which uses service fabric resource provider api-version 2021-05-01.